chore: update seed script
