+++ src/scripts/wipe.ts
@@ -1,12 +1,3 @@
-import db from "@/db";
-import { cardContents, cards, decks, reviewLogs } from "@/schema";
-import { success } from "@/utils/format";
+import { wipeDatabase } from "@/scripts/utils";
 
-export async function wipeDatabase() {
-  console.log("Deleting all data from the database");
-  await db.delete(reviewLogs).all();
-  await db.delete(cardContents).all();
-  await db.delete(cards).all();
-  await db.delete(decks).all();
-  console.log(success`Deleted all data from the database`);
-}
+wipeDatabase();
